double avg volume
closes #169

+++ InterRegionMarketDataModel.h
@@ -94,7 +94,7 @@ namespace Evernus
             double mDstBuyPrice = 0.;
             double mDstSellPrice = 0.;
             double mDifference = 0.;
-            uint mVolume = 0;
+            double mVolume = 0;
             uint mSrcRegion = 0;
             uint mDstRegion = 0;
         };
